requirements.txt
# Core
playwright>=1.40.0
fastapi>=0.104.0
pydantic>=2.0.0
uvicorn>=0.24.0

# LLM & Agents
openai>=1.0.0
anthropic>=0.7.0
langchain>=0.1.0
langchain-openai>=0.0.0
langchain-anthropic>=0.0.0

# Integrations
requests>=2.31.0
pyyaml>=6.0
networkx>=3.2
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Utilities
click>=8.1.0
rich>=13.7.0
tabulate>=0.9.0
python-dotenv>=1.0.0
colorama>=0.4.6

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
